feat: minor improvements and hook file renaming
