+++ packages/app/src/components/form-fields/image/index.tsx
@@ -117,12 +117,10 @@ const Image = ({ name, feedId }: ImageProps) => {
             {!hasRemote && (
               <div>
                 <Link to={`/feed/${feedId}/feed-config`}>
-                  <a>
-                    <Note>
-                      Configure remote storage for uploading files from your
-                      computer
-                    </Note>
-                  </a>
+                  <Note>
+                    Configure remote storage for uploading files from your
+                    computer
+                  </Note>
                 </Link>
               </div>
             )}
