+++ packages/app/src/views/configuration/index.tsx
@@ -7,8 +7,8 @@ import FormRow from "@fourviere/ui/lib/form/form-row";
 import Input from "@fourviere/ui/lib/form/fields/input";
 import { Formik } from "formik";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
-import useConfigurations from "../../hooks/useConfigurations";
-import useTranslations from "../../hooks/useTranslations";
+import useConfigurations from "../../hooks/use-configurations";
+import useTranslations from "../../hooks/use-translations";
 import Select from "@fourviere/ui/lib/form/fields/select";
 import TRANSLATIONS from "../../translations";
 import Boolean from "@fourviere/ui/lib/form/fields/boolean";
