+++ packages/app/src/views/edit-feed/configuration.tsx
@@ -5,8 +5,8 @@ import Input from "@fourviere/ui/lib/form/fields/input";
 import Boolean from "@fourviere/ui/lib/form/fields/boolean";
 import { Formik } from "formik";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
-import UseCurrentFeed from "../../hooks/useCurrentFeed";
-import useTranslations from "../../hooks/useTranslations";
+import UseCurrentFeed from "../../hooks/use-current-feed";
+import useTranslations from "../../hooks/use-translations";
 import Select from "@fourviere/ui/lib/form/fields/select";
 import { FC } from "react";
 import ContainerTitle from "@fourviere/ui/lib/container-title";
