+++ packages/app/src/views/edit-feed/general.tsx
@@ -4,8 +4,8 @@ import FormRow from "@fourviere/ui/lib/form/form-row";
 import Input from "@fourviere/ui/lib/form/fields/input";
 import { Formik } from "formik";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
-import UseCurrentFeed from "../../hooks/useCurrentFeed";
-import useTranslations from "../../hooks/useTranslations";
+import UseCurrentFeed from "../../hooks/use-current-feed";
+import useTranslations from "../../hooks/use-translations";
 import Select from "@fourviere/ui/lib/form/fields/select";
 import PODCASTCATEGORIES from "@fourviere/core/lib/podcast-namespace/categories";
 import { FC } from "react";
