+++ packages/app/src/views/edit-feed/itunes.tsx
@@ -4,8 +4,8 @@ import FormRow from "@fourviere/ui/lib/form/form-row";
 import Input from "@fourviere/ui/lib/form/fields/input";
 import { Formik } from "formik";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
-import UseCurrentFeed from "../../hooks/useCurrentFeed";
-import useTranslations from "../../hooks/useTranslations";
+import UseCurrentFeed from "../../hooks/use-current-feed.tsx";
+import useTranslations from "../../hooks/use-translations.tsx";
 import FormObjectField from "@fourviere/ui/lib/form/form-object-field";
 import Boolean from "@fourviere/ui/lib/form/fields/boolean";
 import { Categories } from "../../components/form-fields/categories.tsx";
