+++ packages/app/src/views/edit-item/item-general.tsx
@@ -6,8 +6,8 @@ import AudioField from "../../components/form-fields/audio";
 import { Formik } from "formik";
 import Img from "../../components/form-fields/image";
 import { FormField } from "@fourviere/ui/lib/form/form-field";
-import UseCurrentFeed from "../../hooks/useCurrentFeed";
-import useTranslations from "../../hooks/useTranslations";
+import UseCurrentFeed from "../../hooks/use-current-feed";
+import useTranslations from "../../hooks/use-translations";
 import { useParams } from "react-router-dom";
 import { FullPageColumnLayout } from "@fourviere/ui/lib/layouts/full-page";
 import { ItemLink } from "../../components/form-fields/item-link";
